server-python/requirements.txt
fastapi
tortoise-orm[aiomysql]
pytest

